+++ src/AceTime.h
@@ -69,7 +69,7 @@
 #include "ace_time/time_period_mutation.h"
 
 // Version format: xxyyzz == "xx.yy.zz"
-#define ACE_TIME_VERSION 10802
-#define ACE_TIME_VERSION_STRING "1.8.2"
+#define ACE_TIME_VERSION 10900
+#define ACE_TIME_VERSION_STRING "1.9.0"
 
 #endif