+++ CHANGELOG.md
@@ -12,6 +12,9 @@
       AceTime compared to Hinnant date library.
         * AceTime seems to be about 90X faster for converting date-time
           components to epoch seconds.
+    * Add `MaxBufSize` comment field into `zonedb[x]/zone_infos.h` which is the
+      maximum over all zones in that file. Must be less than or equal to
+      `ExtendedZoneProcessor::kMaxTransitions`.
 * 1.8.2 (2021-10-28, TZDB 2021e)
     * Update to TZDB 2021e.
         * https://mm.icann.org/pipermail/tz-announce/2021-October/000069.html